Project is now multi module
