+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.2.50'
+    ext.kotlin_version = '1.2.60'
 
     repositories {
         mavenCentral()
@@ -9,37 +9,10 @@ buildscript {
     }
 }
 
-apply plugin: 'kotlin'
-apply plugin: 'maven'
-
-group = 'com.github.fp-in-bo'
-
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-
-    testImplementation 'com.squareup.retrofit2:converter-jackson:2.4.0'
-    testImplementation 'com.squareup.okhttp3:logging-interceptor:3.10.0'
-    testImplementation 'junit:junit:4.12'
-
-
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.6"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.6"
-}
-
-test {
-    useJUnit {
-        includeCategories 'com.fpinbo.kall.category.UnitTest'
-        excludeCategories 'com.fpinbo.kall.category.IntegrationTest'
+allprojects {
+    repositories {
+        mavenCentral()
     }
 }
+
+
