+++ kall/build.gradle
@@ -1,0 +1,26 @@
+apply plugin: 'kotlin'
+group = 'com.github.fp-in-bo'
+
+
+dependencies {
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
+
+    testImplementation 'com.squareup.retrofit2:converter-jackson:2.4.0'
+    testImplementation 'com.squareup.okhttp3:logging-interceptor:3.10.0'
+    testImplementation 'junit:junit:4.12'
+}
+
+test {
+    useJUnit {
+        includeCategories 'com.fpinbo.kall.category.UnitTest'
+        excludeCategories 'com.fpinbo.kall.category.IntegrationTest'
+    }
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.6"
+}
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.6"
+}
