+++ kall/src/test/java/com/fpinbo/kall/TestApplicativeKall.kt
