+++ kall/src/test/java/com/fpinbo/kall/TestFunctorKall.kt
