+++ kall/src/test/java/com/fpinbo/kall/TestMonadKall.kt
