+++ kall/src/test/java/com/fpinbo/kall/api/jokes/Response.kt
