+++ kall/src/test/java/com/fpinbo/kall/response/TestApplicativeResponse.kt
