+++ kall/src/test/java/com/fpinbo/kall/response/TestFunctorResponse.kt
