+++ settings.gradle
@@ -1,2 +1,3 @@
-rootProject.name = 'kall'
+include ':kall'
+
